+++ example/src/index.tsx
@@ -1,10 +1,11 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import HooksForm from './HooksForm'
+import { HooksForm } from './HooksForm'
 
-function App() {
+function App(): JSX.Element {
   return <HooksForm />
 }
 
+// eslint-disable-next-line no-undef
 const rootElement = document.getElementById('root')
 ReactDOM.render(<App />, rootElement)